# Add warning logs for unsupported objects