Prints out error logs when ListObjects return objects which are unsupported i.e. objects with names starting with / or containing // .

Perf test results: [full logs](http://fusion2/ci/kokoro/prod%3Agcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/ae786e77-5c05-41d4-b76d-663a1125eb9e)
```
+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   | 409.24MiB/s  | 1.38MiB/s  |  56.97MiB/s  |  1.32MiB/s   |
|   PR   |  0.25MiB   | 410.66MiB/s  | 1.17MiB/s  |  56.94MiB/s  |  1.29MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 4099.98MiB/s | 78.58MiB/s | 986.55MiB/s  |  77.7MiB/s   |
|   PR   | 48.828MiB  | 4099.1MiB/s  | 77.88MiB/s | 1069.22MiB/s |  77.18MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4109.59MiB/s | 24.92MiB/s | 371.69MiB/s  |  29.17MiB/s  |
|   PR   | 976.562MiB | 4301.21MiB/s | 28.65MiB/s | 443.26MiB/s  |  31.47MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+
```

Proper fix for this is in #1876 .  

### Description

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - NA
   If bucket contains the following objects: `a/b/c`, `a/d//e`, `a/x/y//z`, `a/x/x2`
   Before change, there are no error logs.
   After change, error logs in log file:
   ```
    Encountered unsupported prefixes during listing: [a/d//]
    Encountered unsupported prefixes during listing: [a/x/y//]
   ```
3. Unit tests - NA
4. Integration tests - NA